refactor: split ColumnarStreams into ColumnarInputStream and ColumnarSourceOp
Separate the pull-based and push-based Columnar execution models
into independent files for better maintainability and future pipeline
optimization:

- ColumnarInputStream.h/.cpp: pull-based IProfilingBlockInputStream model (legacy)
- ColumnarSourceOp.h/.cpp: push-based SourceOp pipeline model (future)

This allows pipeline-specific optimizations (async IO, adaptive
concurrency, backpressure) to be scoped to ColumnarSourceOp without
touching the legacy InputStream path.
