+++ dbms/src/Storages/Columnar/ColumnarInputStream.h
@@ -1,4 +1,4 @@
-// Copyright 2026 PingCAP, Inc.
+// Copyright 2023 PingCAP, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,11 +17,9 @@
 #include <Common/config.h> // for ENABLE_NEXT_GEN_COLUMNAR
 #if ENABLE_NEXT_GEN_COLUMNAR
 #include <Common/Logger.h>
-#include <Common/Stopwatch.h>
 #include <DataStreams/AddExtraTableIDColumnTransformAction.h>
 #include <DataStreams/IProfilingBlockInputStream.h>
 #include <Interpreters/Context_fwd.h>
-#include <Operators/Operator.h>
 #include <Storages/Columnar/ColumnarReader.h>
 
 namespace DB
@@ -96,67 +94,6 @@ private:
     double duration_read_sec = 0;
     UInt64 batch_size = 10240;
     UInt64 total_bytes = 0;
-};
-
-class ColumnarSourceOp : public SourceOp
-{
-    static constexpr auto NAME = "RNProxy";
-
-public:
-    struct Options
-    {
-        PipelineExecutorContext & exec_context;
-        ColumnarReadTaskPtr task;
-    };
-
-    explicit ColumnarSourceOp(const Options & options)
-        : SourceOp(options.exec_context, options.task->getLog()->identifier())
-        , context(options.task->getContext())
-        , log(options.task->getLog())
-        , task(options.task)
-    {
-        setHeader(AddExtraTableIDColumnTransformAction::buildHeader(
-            options.task->getColumnsToRead(),
-            options.task->getExtraTableIDIndex()));
-    }
-
-    static SourceOpPtr create(const Options & options) { return std::make_unique<ColumnarSourceOp>(options); }
-
-    String getName() const override { return NAME; }
-
-    IOProfileInfoPtr getIOProfileInfo() const override { return IOProfileInfo::createForLocal(profile_info_ptr); }
-
-protected:
-    void operateSuffixImpl() override;
-
-    void operatePrefixImpl() override;
-
-    OperatorStatus readImpl(Block & block) override;
-
-    OperatorStatus awaitImpl() override;
-
-    OperatorStatus executeIOImpl() override;
-
-private:
-    const Context & context;
-    const LoggerPtr log;
-    ColumnarReadTaskPtr task;
-    UInt64 total_bytes = 0;
-    size_t total_rows = 0;
-    size_t total_streams = 0;
-
-    BlockInputStreamPtr current_input_stream;
-
-    // Temporarily store the block read from current_seg_task->stream and pass it to downstream operators in readImpl.
-    std::optional<Block> t_block = std::nullopt;
-
-    bool done = false;
-    // Count the time spent waiting for segment tasks to be ready.
-    //double duration_wait_ready_task_sec = 0;
-    Stopwatch total_cost_watch{CLOCK_MONOTONIC_COARSE};
-
-    // Count the time consumed by reading blocks in the stream of segment tasks.
-    double duration_read_sec = 0;
 };
 
 } // namespace DB
